feat(ecg-chart): create new folders and separate DataContext into DataContext and useDataContext files
